pyproject.toml
[tool.poetry]
name = "docker_study_project"
version = "0.1.0"
description = "Simple project with docker, fastapi unicovn, postgree"
authors = ["Vladimir Kondakov <vladimir.kondakov@me.com>"]

[tool.poetry.dependencies]
python = "3.9"
fastapi = "^0.70.1"
uvicorn = "^0.16.0"
SQLAlchemy = "^1.4.29"
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.2"
python-dotenv = "^0.19.2"
email-validator = "^1.1.3"
loguru = "^0.5.3"
PyJWT = "^2.3.0"
passlib = "^1.7.4"
bcrypt = "^3.2.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
autopep8 = "^1.6.0"
httpx = "^0.21.1"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
